ai-assistant/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
llama-index==0.10.12
llama-index-vector-stores-weaviate==1.0.0
llama-index-llms-openai==0.1.6
llama-index-embeddings-openai==0.1.6
weaviate-client==4.4.4
confluent-kafka==2.3.0
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
python-dotenv==1.0.1
